+++ src/io/output_stream.ts
@@ -24,3 +24,5 @@ export class BufferedRenderStream implements RenderStream {
     return this.buffer.join("");
   }
 }
+
+// TODO: A RenderStream with resource limits?